+++ beatmap/v3/parse.ts
@@ -3,7 +3,6 @@ import { Difficulty } from './difficulty.ts';
 import { DifficultyCheck } from './dataCheck.ts';
 import { deepCheck } from '../shared/dataCheck.ts';
 import logger from '../../logger.ts';
-import { sortV3NoteFn, sortV3ObjectFn } from '../shared/helpers.ts';
 import { IDataCheckOption } from '../../types/beatmap/shared/dataCheck.ts';
 
 function tag(name: string): string[] {
@@ -45,20 +44,6 @@ export function parseDifficulty(
       e.w ??= 0;
       e.h ??= 0;
    });
-
-   data.bpmEvents?.sort(sortV3ObjectFn);
-   data.rotationEvents?.sort(sortV3ObjectFn);
-   data.colorNotes?.sort(sortV3NoteFn);
-   data.bombNotes?.sort(sortV3NoteFn);
-   data.obstacles?.sort(sortV3ObjectFn);
-   data.sliders?.sort(sortV3NoteFn);
-   data.burstSliders?.sort(sortV3NoteFn);
-   data.waypoints?.sort(sortV3ObjectFn);
-   data.basicBeatmapEvents?.sort(sortV3ObjectFn);
-   data.colorBoostBeatmapEvents?.sort(sortV3ObjectFn);
-   data.lightColorEventBoxGroups?.sort(sortV3ObjectFn);
-   data.lightRotationEventBoxGroups?.sort(sortV3ObjectFn);
-   data.lightTranslationEventBoxGroups?.sort(sortV3ObjectFn);
 
    return new Difficulty(data);
 }