+++ CHANGELOG.md
@@ -27,6 +27,8 @@ faster in general use case.
 ### Changed
 
 - Beatmap v3 now purge all zero value attributes on optimized save
+- Parsing no longer handle sorting
+  - Moved to class method, sorting is done by default on load
 - The Weeknd and Panic 2.0 color scheme now shows white color value
 - Utils no longer require namespace
   - This should reveal more hidden utilities without the need to import and destructuring `utils`